completed the HibernateSessions wrapper and added a lot of error checking
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@911 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/dbsupport/HibernateDataProviderSession.java
@@ -4,13 +4,11 @@
 * [INSERT KU-APPROVED LICENSE TEXT HERE]
 *
 */
-/**
- * 
- */
 package edu.ku.brc.specify.dbsupport;
 
 import java.util.List;
 
+import org.apache.log4j.Logger;
 import org.hibernate.Criteria;
 import org.hibernate.LockMode;
 import org.hibernate.Query;
@@ -25,23 +23,34 @@ import edu.ku.brc.dbsupport.HibernateUtil;
 import edu.ku.brc.dbsupport.StaleObjectException;
 
 /**
+ * This is a wrapper around Hibernate Session so we don't have to pollute our class with Hibernate and we could switch it
+ * out for a networked version later.
+ * 
  * @author rods
  *
- * @code_status Alpha
+ * @code_status Complete
  *
  */
 public class HibernateDataProviderSession implements DataProviderSessionIFace
 {
-
+    private static final Logger log = Logger.getLogger(HibernateDataProviderSession.class);
+    
+    // Used for checking to see if we have any dangling creates without closes
+    //protected static      int creates = 0;
+    //protected static      int closes  = 0;
+    
     protected Session     session         = null;
     protected Exception   recentException = null;
     protected Transaction transaction     = null;
+    
     /**
-     * 
+     * Creates a new Hibernate Session
      */
     public HibernateDataProviderSession()
     {
         session = HibernateUtil.getNewSession();
+        //creates++;
+        //log.info(" Creates: "+creates+"  Closes: "+closes+" Dif: "+(creates-closes));
     }
 
     /* (non-Javadoc)
@@ -49,11 +58,15 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public boolean delete(Object dataObj) throws Exception
     {
+        if (session != null)
+        {
+            session.delete(dataObj);
+            return false;
+        }
+        
+        log.error("Session was null.");
 
-        session.delete(dataObj);
-            
-
-        return true;
+        return false;
     }
 
     /* (non-Javadoc)
@@ -61,8 +74,16 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public List getDataList(String sqlStr)
     {
-        Query query = session.createQuery(sqlStr);
-        return query.list();
+        if (session != null)
+        {
+            Query query = session.createQuery(sqlStr);
+            return query.list();
+        }
+        
+        log.error("Session was null.");
+
+        return null;
+
     }
     
     /* (non-Javadoc)
@@ -70,8 +91,15 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public List getDataList(Class clsObject)
     {
-        Criteria criteria = session.createCriteria(clsObject);
-        return criteria.list();
+        if (session != null)
+        {
+            Criteria criteria = session.createCriteria(clsObject);
+            return criteria.list();
+        }
+        
+        log.error("Session was null.");
+
+        return null;
     }
     
     /* (non-Javadoc)
@@ -79,9 +107,16 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public List getDataList(Class clsObject, String fieldName, Object value, DataProviderSessionIFace.CompareType compareType)
     {
-        Criteria criteria = session.createCriteria(clsObject);
-        criteria.add(compareType == DataProviderSessionIFace.CompareType.Equals ? Expression.eq(fieldName, value) : Restrictions.eq(fieldName, value));
-        return criteria.list();
+        if (session != null)
+        {
+            Criteria criteria = session.createCriteria(clsObject);
+            criteria.add(compareType == DataProviderSessionIFace.CompareType.Equals ? Expression.eq(fieldName, value) : Restrictions.eq(fieldName, value));
+            return criteria.list();           
+        }
+        
+        log.error("Session was null.");
+
+        return null;
     }
     
     /* (non-Javadoc)
@@ -89,7 +124,14 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public List getDataList(Class clsObject, String fieldName, Object value)
     {
-        return getDataList(clsObject, fieldName, value, DataProviderSessionIFace.CompareType.Equals);
+        if (session != null)
+        {
+            return getDataList(clsObject, fieldName, value, DataProviderSessionIFace.CompareType.Equals);
+        }
+        
+        log.error("Session was null.");
+
+        return null;
     }
     
     /* (non-Javadoc)
@@ -105,7 +147,14 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public Object load(Class clsObj, Long id)
     {
-        return session.load(clsObj, id);
+        if (session != null)
+        {
+            return session.load(clsObj, id);
+        }
+        
+        log.error("Session was null.");
+
+        return null;
     }
     
     /* (non-Javadoc)
@@ -113,8 +162,15 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public Object getData(String sqlStr)
     {
-        List list = getDataList(sqlStr);
-        return list != null && list.size() > 0 ? list.get(0) : null;
+        if (session != null)
+        {
+            List list = getDataList(sqlStr);
+            return list != null && list.size() > 0 ? list.get(0) : null;
+        }
+        
+        log.error("Session was null.");
+
+        return null;
     }
 
     /* (non-Javadoc)
@@ -122,7 +178,13 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public void evict(Class clsObject)
     {
-        session.evict(clsObject);
+        if (session != null)
+        {
+            session.evict(clsObject);
+        } else
+        {
+            log.error("Session was null.");
+        }
     }
     
     /* (non-Javadoc)
@@ -130,8 +192,13 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public void evict(Object dataObj)
     {
-        session.evict(dataObj);
-       
+        if (session != null)
+        {
+            session.evict(dataObj);
+        } else
+        {
+            log.error("Session was null.");
+        }
     }
     
     /* (non-Javadoc)
@@ -139,7 +206,14 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public void attach(Object dataObj)
     {
-        session.lock(dataObj, LockMode.NONE);
+        if (session != null)
+        {
+            session.lock(dataObj, LockMode.NONE);
+            
+        } else
+        {
+            log.error("Session was null.");
+        }
     }
     
     /* (non-Javadoc)
@@ -147,9 +221,15 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public boolean save(Object dataObj) throws Exception
     {
-        session.save(dataObj);
+        if (session != null)
+        {
+            session.save(dataObj);
+            return true;
+        }
         
-        return true;
+        log.error("Session was null.");
+        
+        return false;
     }
 
     /* (non-Javadoc)
@@ -157,9 +237,15 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public boolean saveOrUpdate(Object dataObj) throws Exception
     {
-        session.saveOrUpdate(dataObj);
+        if (session != null)
+        {
+            session.saveOrUpdate(dataObj);
+            return true;
+        }
         
-        return true;
+        log.error("Session was null.");
+        
+        return false;
     }
 
     /* (non-Javadoc)
@@ -167,9 +253,15 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public boolean update(Object dataObj) throws Exception
     {
-        session.saveOrUpdate(dataObj);
+        if (session != null)
+        {
+            session.update(dataObj);
+            return true;
+        }
         
-        return true;
+        log.error("Session was null.");
+        
+        return false;
     }
     
     /* (non-Javadoc)
@@ -177,7 +269,17 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public void beginTransaction() throws Exception
     {
-        transaction = session.beginTransaction();
+        if (session != null)
+        {
+            transaction = session.beginTransaction();
+            if (transaction == null)
+            {
+                log.error("Transaction was null!"); // Throw Exception?
+            }
+        } else
+        {
+            log.error("Session was null.");
+        }
     }
     
     /* (non-Javadoc)
@@ -196,6 +298,9 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
             {
                 throw new StaleObjectException(soe);
             }
+        } else
+        {
+            throw new RuntimeException("Transaction was null and shouldn't been.");
         }
     }
     
@@ -207,6 +312,10 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
         if (transaction != null)
         {
             transaction.rollback();
+            
+        } else
+        {
+            throw new RuntimeException("Transaction was null and shouldn't been.");
         }
     }
 
@@ -215,7 +324,14 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public void flush()
     {
-        session.flush();
+        if (session != null)
+        {
+            session.flush();
+            
+        } else
+        {
+            log.error("Session was null.");
+        }
     }
     
     /* (non-Javadoc)
@@ -223,8 +339,23 @@ public class HibernateDataProviderSession implements DataProviderSessionIFace
      */
     public void close()
     {
-        session.close();
-        session     = null;
-        transaction = null;
-    }
+        //closes++;
+        //log.info("*Creates: "+creates+"  Closes: "+closes+" Dif: "+(creates-closes));
+        
+        if (session != null)
+        {
+            if (transaction != null)
+            {
+                transaction.rollback();
+                transaction = null;
+                throw new RuntimeException("Closing Session with open transaction - rolling it back");
+            }
+            
+            session.close();
+            session = null;
+        } else
+        {
+            log.error("Session was null.");
+        }
+     }
 }
